Support for storing multi-frame images* (#22)
* #11 Refactor: Paved way to start developing support for multi-frame dicom objects

* #11 Refactor: Memory optimization when codec -> all

Optimization is to use iterator, for each dicom object clone, instead of array - more memory efficient
Also, fixed bug when getting uri (was not getting the correct TS sometimes)
Also, changed some dependencies

* #11 Feature: Store multi-frame images

* #11 Finished feature*: Viewing raw multi-frame images as gif

 * Workaround - Could not find a way yet to display animated images. Thus, server can only view them locally by a file:// link
 * I marked the source of the above problem with a 'fixme 16-09-22' tag

* #11 Finished feature: Multi-frame store operation

* #11 [Bugfix] Some old features that were broken

By the multi-frame support feature

* Code cleanup

* [Refactor] Moves functions pertaining utility to *Utils.java

* #11 [Bugfix] could not read single frame avif

* #11 [README] Informed that `all` does not work with multi-frames storing operation

+++ src/main/java/pt/ua/imodec/util/NewFormatsCodecs.java
@@ -78,7 +78,7 @@ public class NewFormatsCodecs {
                     return String.format("cavif -o %s --quality %s --speed %s %s", outputPath, quality, speed,
                             inputPath);
                 }
-                return String.format("avif_decode %s %s", inputPath, outputPath);
+                return String.format("avif_decode -f %s %s", inputPath, outputPath);
             case "webp":
                 if (encoding) {
                     Number quality = options.getOrDefault("quality", NewFormat.WEBP.getQualityParamValue()),
@@ -93,7 +93,7 @@ public class NewFormatsCodecs {
     }
 
     public static BufferedImage decodeByteStream(byte[] bitstream, NewFormat chosenFormat) throws IOException {
-        String encodedFileName = String.format("%s/%s.%s", ImodecPluginSet.tmpDirPath,
+        String encodedFileName = String.format("%s/%s.%s", ImodecPluginSet.TMP_DIR_PATH,
                 Arrays.hashCode(bitstream), chosenFormat.getFileExtension());
         Files.write(Paths.get(encodedFileName), bitstream);
         return decode(encodedFileName, chosenFormat.getFileExtension());
@@ -102,13 +102,15 @@ public class NewFormatsCodecs {
     private static BufferedImage decode(String inputFilePath, String formatExtension) throws IOException {
 
         String decodedFileName = inputFilePath.replace(formatExtension, losslessFormat);
+        File decodedImageFile = new File(decodedFileName);
+        if (decodedImageFile.exists())
+            return ImageIO.read(decodedImageFile);
 
         String decodingCommand = getCodecCommand(inputFilePath, decodedFileName, formatExtension, false,
                 new HashMap<>());
 
         execute(decodingCommand);
 
-        File decodedImageFile = new File(decodedFileName);
         decodedImageFile.deleteOnExit();
 
         return ImageIO.read(decodedImageFile);
